use unified Jni Instance
