+++ jvm/src/main/java/com/intel/oap/vectorized/ShuffleSplitterJniWrapper.java
@@ -22,7 +22,7 @@ import java.io.IOException;
 public class ShuffleSplitterJniWrapper {
 
   public ShuffleSplitterJniWrapper() throws IOException {
-    JniUtils.getInstance();
+    JniInstance.getInstance();
   }
 
   /**
